// src/app/models/login.model.ts
export interface LoginRequest {
  username?: string;
  password?: string;
}
export interface LoginResponse {
  id: number;
  username?: string;
  email?: string;
  firstName?: string;
  lastName?: string;
  gender?: string;
  image?: string;
  accessToken: string;
  refreshToken?: string;
  error: number;
}